Refactoring: Extract user event handling into UserEventHandler type
